Module aws_sdk_inspector::error

Modules

access_denied_exception

See AccessDeniedException

agents_already_running_assessment_exception

See AgentsAlreadyRunningAssessmentException

assessment_run_in_progress_exception

See AssessmentRunInProgressException

internal_exception

See InternalException

invalid_cross_account_role_exception

See InvalidCrossAccountRoleException

invalid_input_exception

See InvalidInputException

limit_exceeded_exception

See LimitExceededException

no_such_entity_exception

See NoSuchEntityException

preview_generation_in_progress_exception

See PreviewGenerationInProgressException

service_temporarily_unavailable_exception

See ServiceTemporarilyUnavailableException

unsupported_feature_exception

See UnsupportedFeatureException

Structs

AccessDeniedException

You do not have required permissions to access the requested resource.

AddAttributesToFindingsError

AgentsAlreadyRunningAssessmentException

You started an assessment run, but one of the instances is already participating in

AssessmentRunInProgressException

You cannot perform a specified action if an assessment run is currently in

CreateAssessmentTargetError

CreateAssessmentTemplateError

CreateExclusionsPreviewError

CreateResourceGroupError

DeleteAssessmentRunError

DeleteAssessmentTargetError

DeleteAssessmentTemplateError

DescribeAssessmentRunsError

DescribeAssessmentTargetsError

DescribeAssessmentTemplatesError

DescribeCrossAccountAccessRoleError

DescribeExclusionsError

DescribeFindingsError

DescribeResourceGroupsError

DescribeRulesPackagesError

GetAssessmentReportError

GetExclusionsPreviewError

GetTelemetryMetadataError

InternalException

Internal server error.

InvalidCrossAccountRoleException

Amazon Inspector cannot assume the cross-account role that it needs to list your EC2

InvalidInputException

The request was rejected because an invalid or out-of-range value was supplied for an

LimitExceededException

The request was rejected because it attempted to create resources beyond the current

ListAssessmentRunAgentsError

ListAssessmentRunsError

ListAssessmentTargetsError

ListAssessmentTemplatesError

ListEventSubscriptionsError

ListExclusionsError

ListFindingsError

ListRulesPackagesError

ListTagsForResourceError

NoSuchEntityException

The request was rejected because it referenced an entity that does not exist. The

PreviewAgentsError

PreviewGenerationInProgressException

The request is rejected. The specified assessment template is currently generating an

RegisterCrossAccountAccessRoleError

RemoveAttributesFromFindingsError

ServiceTemporarilyUnavailableException

The serice is temporary unavailable.

SetTagsForResourceError

StartAssessmentRunError

StopAssessmentRunError

SubscribeToEventError

UnsubscribeFromEventError

UnsupportedFeatureException

Used by the GetAssessmentReport API. The request was rejected

UpdateAssessmentTargetError

Enums

AddAttributesToFindingsErrorKind

CreateAssessmentTargetErrorKind

CreateAssessmentTemplateErrorKind

CreateExclusionsPreviewErrorKind

CreateResourceGroupErrorKind

DeleteAssessmentRunErrorKind

DeleteAssessmentTargetErrorKind

DeleteAssessmentTemplateErrorKind

DescribeAssessmentRunsErrorKind

DescribeAssessmentTargetsErrorKind

DescribeAssessmentTemplatesErrorKind

DescribeCrossAccountAccessRoleErrorKind

DescribeExclusionsErrorKind

DescribeFindingsErrorKind

DescribeResourceGroupsErrorKind

DescribeRulesPackagesErrorKind

GetAssessmentReportErrorKind

GetExclusionsPreviewErrorKind

GetTelemetryMetadataErrorKind

ListAssessmentRunAgentsErrorKind

ListAssessmentRunsErrorKind

ListAssessmentTargetsErrorKind

ListAssessmentTemplatesErrorKind

ListEventSubscriptionsErrorKind

ListExclusionsErrorKind

ListFindingsErrorKind

ListRulesPackagesErrorKind

ListTagsForResourceErrorKind

PreviewAgentsErrorKind

RegisterCrossAccountAccessRoleErrorKind

RemoveAttributesFromFindingsErrorKind

SetTagsForResourceErrorKind

StartAssessmentRunErrorKind

StopAssessmentRunErrorKind

SubscribeToEventErrorKind

UnsubscribeFromEventErrorKind

UpdateAssessmentTargetErrorKind