Module aws_sdk_inspector::model

Modules

agent_already_running_assessment

See AgentAlreadyRunningAssessment

agent_filter

See AgentFilter

agent_preview

See AgentPreview

assessment_run

See AssessmentRun

assessment_run_agent

See AssessmentRunAgent

assessment_run_filter

See AssessmentRunFilter

assessment_run_notification

See AssessmentRunNotification

assessment_run_state_change

See AssessmentRunStateChange

assessment_target

See AssessmentTarget

assessment_target_filter

See AssessmentTargetFilter

assessment_template

See AssessmentTemplate

assessment_template_filter

See AssessmentTemplateFilter

asset_attributes

See AssetAttributes

attribute

See Attribute

duration_range

See DurationRange

event_subscription

See EventSubscription

exclusion

See Exclusion

exclusion_preview

See ExclusionPreview

failed_item_details

See FailedItemDetails

finding

See Finding

finding_filter

See FindingFilter

inspector_service_attributes

See InspectorServiceAttributes

network_interface

See NetworkInterface

private_ip

See PrivateIp

resource_group

See ResourceGroup

resource_group_tag

See ResourceGroupTag

rules_package

See RulesPackage

scope

See Scope

security_group

See SecurityGroup

subscription

See Subscription

tag

See Tag

telemetry_metadata

See TelemetryMetadata

timestamp_range

See TimestampRange

Structs

AgentAlreadyRunningAssessment

Used in the exception error that is thrown if you start an assessment run for an

AgentFilter

Contains information about an Amazon Inspector agent. This data type is used as a

AgentPreview

Used as a response element in the PreviewAgents action.

AssessmentRun

A snapshot of an Amazon Inspector assessment run that contains the findings of the

AssessmentRunAgent

Contains information about an Amazon Inspector agent. This data type is used as a

AssessmentRunFilter

Used as the request parameter in the ListAssessmentRuns

AssessmentRunNotification

Used as one of the elements of the AssessmentRun data

AssessmentRunStateChange

Used as one of the elements of the AssessmentRun data

AssessmentTarget

Contains information about an Amazon Inspector application. This data type is used as

AssessmentTargetFilter

Used as the request parameter in the ListAssessmentTargets

AssessmentTemplate

Contains information about an Amazon Inspector assessment template. This data type is

AssessmentTemplateFilter

Used as the request parameter in the ListAssessmentTemplates

AssetAttributes

A collection of attributes of the host from which the finding is generated.

Attribute

This data type is used as a request parameter in the AddAttributesToFindings and CreateAssessmentTemplate

DurationRange

This data type is used in the AssessmentTemplateFilter data

EventSubscription

This data type is used in the Subscription data type.

Exclusion

Contains information about what was excluded from an assessment run.

ExclusionPreview

Contains information about what is excluded from an assessment run given the current

FailedItemDetails

Includes details about the failed items.

Finding

Contains information about an Amazon Inspector finding. This data type is used as the

FindingFilter

This data type is used as a request parameter in the ListFindings

InspectorServiceAttributes

This data type is used in the Finding data type.

NetworkInterface

Contains information about the network interfaces interacting with an EC2 instance.

PrivateIp

Contains information about a private IP address associated with a network interface. This

ResourceGroup

Contains information about a resource group. The resource group defines a set of tags

ResourceGroupTag

This data type is used as one of the elements of the ResourceGroup

RulesPackage

Contains information about an Amazon Inspector rules package. This data type is used

Scope

This data type contains key-value pairs that identify various Amazon

SecurityGroup

Contains information about a security group associated with a network interface. This

Subscription

This data type is used as a response element in the ListEventSubscriptions action.

Tag

A key and value pair. This data type is used as a request parameter in the SetTagsForResource action and a response element in the ListTagsForResource action.

TelemetryMetadata

The metadata about the Amazon Inspector application data metrics collected by the

TimestampRange

This data type is used in the AssessmentRunFilter data

Enums

AccessDeniedErrorCode

AgentHealth

NOTE: AgentHealth::Unknown has been renamed to ::UnknownValue.

AgentHealthCode

NOTE: AgentHealthCode::Unknown has been renamed to ::UnknownValue.

AssessmentRunNotificationSnsStatusCode

AssessmentRunState

AssetType

FailedItemErrorCode

InspectorEvent

InvalidCrossAccountRoleErrorCode

InvalidInputErrorCode

LimitExceededErrorCode

Locale

NoSuchEntityErrorCode

PreviewStatus

ReportFileFormat

ReportStatus

ReportType

ScopeType

Severity

StopAction