Module aws_sdk_inspector::operation

Structs

AddAttributesToFindings

Assigns attributes (key and value pairs) to the findings that are specified by the

CreateAssessmentTarget

Creates a new assessment target using the ARN of the resource group that is generated

CreateAssessmentTemplate

Creates an assessment template for the assessment target that is specified by the ARN

CreateExclusionsPreview

Starts the generation of an exclusions preview for the specified assessment template.

CreateResourceGroup

Creates a resource group using the specified set of tags (key and value pairs) that

DeleteAssessmentRun

Deletes the assessment run that is specified by the ARN of the assessment

DeleteAssessmentTarget

Deletes the assessment target that is specified by the ARN of the assessment

DeleteAssessmentTemplate

Deletes the assessment template that is specified by the ARN of the assessment

DescribeAssessmentRuns

Describes the assessment runs that are specified by the ARNs of the assessment

DescribeAssessmentTargets

Describes the assessment targets that are specified by the ARNs of the assessment

DescribeAssessmentTemplates

Describes the assessment templates that are specified by the ARNs of the assessment

DescribeCrossAccountAccessRole

Describes the IAM role that enables Amazon Inspector to access your AWS

DescribeExclusions

Describes the exclusions that are specified by the exclusions' ARNs.

DescribeFindings

Describes the findings that are specified by the ARNs of the findings.

DescribeResourceGroups

Describes the resource groups that are specified by the ARNs of the resource

DescribeRulesPackages

Describes the rules packages that are specified by the ARNs of the rules

GetAssessmentReport

Produces an assessment report that includes detailed and comprehensive results of a

GetExclusionsPreview

Retrieves the exclusions preview (a list of ExclusionPreview objects) specified by

GetTelemetryMetadata

Information about the data that is collected for the specified assessment

ListAssessmentRunAgents

Lists the agents of the assessment runs that are specified by the ARNs of the

ListAssessmentRuns

Lists the assessment runs that correspond to the assessment templates that are

ListAssessmentTargets

Lists the ARNs of the assessment targets within this AWS account. For more

ListAssessmentTemplates

Lists the assessment templates that correspond to the assessment targets that are

ListEventSubscriptions

Lists all the event subscriptions for the assessment template that is specified by

ListExclusions

List exclusions that are generated by the assessment run.

ListFindings

Lists findings that are generated by the assessment runs that are specified by the

ListRulesPackages

Lists all available Amazon Inspector rules packages.

ListTagsForResource

Lists all tags associated with an assessment template.

PreviewAgents

Previews the agents installed on the EC2 instances that are part of the specified

RegisterCrossAccountAccessRole

Registers the IAM role that grants Amazon Inspector access to AWS Services needed to

RemoveAttributesFromFindings

Removes entire attributes (key and value pairs) from the findings that are specified

SetTagsForResource

Sets tags (key and value pairs) to the assessment template that is specified by the

StartAssessmentRun

Starts the assessment run specified by the ARN of the assessment template. For this

StopAssessmentRun

Stops the assessment run that is specified by the ARN of the assessment

SubscribeToEvent

Enables the process of sending Amazon Simple Notification Service (SNS) notifications

UnsubscribeFromEvent

Disables the process of sending Amazon Simple Notification Service (SNS)

UpdateAssessmentTarget

Updates the assessment target that is specified by the ARN of the assessment